requirements/dev.txt
#requirements/dev.txt
-r common.txt
django-debug-toolbar==0.9.4

